tweaks for compiling patterns
